src/templates/backend/fastapi/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
{{#if backend.eslint}}
pylint>=3.0.0
{{/if}}
{{#if backend.prettier}}
black>=23.0.0
{{/if}}
